# Taking degeneracies seriously