--- a/Project.toml
+++ b/Project.toml
@@ -4,7 +4,6 @@ authors = ["Pablo San-Jose"]
 version = "0.4.0"
 
 [deps]
-DualNumbers = "fa6b7ba4-c1ee-5f82-b5fc-ecf0adba8f74"
 ExprTools = "e2ba6199-217a-4e67-a87a-7c52f15ade04"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
@@ -20,7 +19,6 @@ StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 
 [compat]
-DualNumbers = "0.6"
 ExprTools = "^0.1"
 GeometryBasics = "^0.3"
 LinearMaps = "2.6"